google-authentication
Created the authentication part for the website. Only allowed user is Lalith karthik's mail id as of now I think, not sure, but mostly yeah. But even without logging in or signing up, the user may go through all functionalities and pages of the website, not exactly as per requirement, but as of now this is the progress.